+++ proxy/src/auth/backend.rs
@@ -11,7 +11,7 @@ use crate::auth::validate_password_and_exchange;
 use crate::console::errors::GetAuthInfoError;
 use crate::console::provider::AuthInfo;
 use crate::console::AuthSecret;
-use crate::proxy::{handle_try_wake, retry_after, LatencyTimer};
+use crate::proxy::{handle_try_wake, retry::retry_after, LatencyTimer};
 use crate::scram;
 use crate::stream::Stream;
 use crate::{
